kernel/futex: compare FUTEX_WAIT/CMP_REQUEUE value as 32-bit (fix int sign-extension EAGAIN storm)
Per futex(2), "the futex word is a 32-bit field" and the FUTEX_WAIT
comparison `*uaddr == val` is defined over those 32 bits only. Our
check-then-queue helper compared the full 64-bit `val` syscall argument
against the zero-extended 32-bit `*uaddr`, which breaks for any futex
word with bit 31 set.

The futex `val` argument is typed `int` in the C-library wrappers
(pthread_mutex_timedlock -> __timedwait(addr, val, ...)). When a thread
waits on a contended mutex, musl passes the mutex word with the
waiters bit (0x8000_0010) as that `int`. Widening a negative `int` to
the 64-bit syscall register sign-extends it, so the kernel receives
0xFFFF_FFFF_8000_0010 for a futex word of 0x8000_0010. The full-width
compare `current as u64 != val` then mismatched in the high half on
every call and returned EAGAIN, so a waiter on a contended mutex spun
instead of parking — observed as a multi-billion-call FUTEX_WAIT storm
with a frozen futex word, and a main thread that never slept to receive
the owner's release wake.

Fix: truncate `val` to 32 bits before the compare (`current != val as u32`),
matching the 32-bit futex-word contract. The same defect existed in the
FUTEX_CMP_REQUEUE `*uaddr == val3` check (pthread_cond broadcast path);
fixed there too.

Verified against a real workload that drove a contended-mutex wait whose
word's bit 31 was set: pre-fix the syscall count climbed past 4x10^8 with
the futex word frozen and no waiter ever enqueued; post-fix the same
workload's syscall count stays bounded (max ~2.8x10^3), waiters enqueue
and park correctly, and execution proceeds past the spin. Reproduced
across two runs with identical outcome.

Adds a deterministic test_runner.rs regression test
(test_futex_wait_val_is_32bit) that drives futex_wait_check_and_enqueue
with a sign-extended `val` and asserts: low-32-bits-agree -> Enqueued,
low-32-bits-differ -> ValueMismatch.

Refs:
  * futex(2) — FUTEX_WAIT / FUTEX_CMP_REQUEUE value comparison, 32-bit
    futex word.
  * POSIX-1.2017 pthread_mutex_lock(3) / pthread_mutex_timedlock(3).

+++ kernel/src/subsys/linux/syscall.rs
@@ -8877,7 +8877,14 @@ pub fn sys_futex_linux(
                     Some(v) => v,
                     None => return -14, // EFAULT
                 };
-                if current as u64 != _val3 {
+                // Per futex(2), the futex word is 32-bit; compare only the low
+                // 32 bits of `val3`.  The `val3` argument is typed `int` in the
+                // C-library wrappers, so a value with bit 31 set arrives in the
+                // 64-bit syscall register sign-extended (high half all-ones).
+                // A full-width compare against the zero-extended 32-bit `*uaddr`
+                // would mismatch in the high half and spuriously return EAGAIN,
+                // breaking the pthread_cond broadcast/requeue handshake.
+                if current != (_val3 as u32) {
                     return -11; // EAGAIN — *uaddr changed under us
                 }
             }
